+++ tests/bootstrap.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * PHP library for save CSS and JS files to be displayed in same place.
+ * 
+ * @author     Josantonius - hello@josantonius.com
+ * @copyright  Copyright (c) 2016 - 2017
+ * @license    https://opensource.org/licenses/MIT - The MIT License (MIT)
+ * @link       https://github.com/Josantonius/PHP-ErrorHandler
+ * @since      1.1.4
+ */
+
+$loader = require __DIR__ . '/../src/bootstrap.php';
+
+$loader->add('Josantonius\ErrorHandler\Test', __DIR__);