+++ CHANGELOG.md
@@ -1,11 +1,52 @@
 # CHANGELOG
 
+## 1.1.5 - 2017-09-13
+
+* Unit tests supported by `PHPUnit` were added.
+
+* The repository was synchronized with Travis CI to implement continuous integration.
+ 
+* Added `ErrorHandler/src/bootstrap.php` file
+
+* Added `ErrorHandler/tests/bootstrap.php` file.
+
+* Added `ErrorHandler/phpunit.xml.dist` file.
+* Added `ErrorHandler/_config.yml` file.
+* Added `ErrorHandler/.travis.yml` file.
+
+* Added `ErrorHandler/public/template/view.php` file.
+
+* Added `ErrorHandler/public/css/styles.php` file.
+
+* Deleted `Josantonius\ErrorHandler\Tests\ErrorHandlerTest` class.
+* Deleted `Josantonius\ErrorHandler\Tests\ErrorHandlerTest->testSProvokeException()` method.
+* Deleted `Josantonius\ErrorHandler\Tests\ErrorHandlerTest->testSProvokeWarning()` method.
+* Deleted `Josantonius\ErrorHandler\Tests\ErrorHandlerTest->testSProvokeNotice()` method.
+* Deleted `Josantonius\ErrorHandler\Tests\ErrorHandlerTest->testSProvokeUserError()` method.
+* Deleted `Josantonius\ErrorHandler\Tests\ErrorHandlerTest->testSProvokeUserNotice()` method.
+* Deleted `Josantonius\ErrorHandler\Tests\ErrorHandlerTest->testSProvokeUserWarning()` method.
+* Deleted `Josantonius\ErrorHandler\Tests\ErrorHandlerTest->testSetCustomMethod()` method.
+* Deleted `Josantonius\ErrorHandler\Tests\ErrorHandlerTest->_customMethodResponse()` method.
+
+* Added `Josantonius\ErrorHandler\Test\ErrorHandlerTest` class.
+* Added `Josantonius\ErrorHandler\Test\ErrorHandlerTest->testIfCustomExceptionHandlerIsActivated()` method.
+* Added `Josantonius\ErrorHandler\Test\ErrorHandlerTest->testIfCustomErrorHandlerIsActivated()` method.
+* Added `Josantonius\ErrorHandler\Test\ErrorHandlerTest->testException()` method.
+* Added `Josantonius\ErrorHandler\Test\ErrorHandlerTest->testExceptionWithCustomMethodsWithoutShowDefaultView()` method.
+* Added `Josantonius\ErrorHandler\Test\ErrorHandlerTest->testExceptionWithCustomMethodsAndShowDefaultView()` method.
+* Added `Josantonius\ErrorHandler\Test\ErrorHandlerTest->testError()` method.
+* Added `Josantonius\ErrorHandler\Test\ErrorHandlerTest->testErrorWithCustomMethodsWithoutShowDefaultView()` method.
+* Added `Josantonius\ErrorHandler\Tests\ErrorHandlerTest->testErrorWithCustomMethodsAndShowDefaultView()` method.
+* Added `Josantonius\ErrorHandler\Tests\ErrorHandlerTest->testGetErrorType()` method.
+
 ## 1.1.4 - 2017-07-16
+
 * Deleted `Josantonius\ErrorHandler\Exception\ErrorHandlerException` class.
 * Deleted `Josantonius\ErrorHandler\Exception\Exceptions` abstract class.
 * Deleted `Josantonius\ErrorHandler\Exception\ErrorHandlerException->__construct()` method.
 
 ## 1.1.3 - 2017-05-15
+
 * You can now add custom methods to run instead of displaying the default view.
 
 * Added `Josantonius\ErrorHandler\ErrorHandler->setParams()` method.
@@ -16,9 +57,11 @@
 * Added `Josantonius\ErrorHandler\Tests\ErrorHandlerTest->_customMethodResponse()` method.
 
 ## 1.1.2 - 2017-03-18
+
 * Some files were excluded from download and comments and readme files were updated.
 
 ## 1.1.1 - 2017-02-21
+
 * Added `Josantonius\ErrorHandler\ErrorHandler->getPreviewCode()` method.
 * Deleted `Josantonius\ErrorHandler\ErrorHandler->catchThrowable()` method.
 * Deleted `Josantonius\ErrorHandler\ErrorHandler->catchException()` method.
@@ -27,12 +70,15 @@
 * Changed `Josantonius\ErrorHandler\ErrorHandler->show()` method to `Josantonius\ErrorHandler\ErrorHandler->render()`.
 
 ## 1.1.0 - 2017-01-30
+
 * Compatible with PHP 5.6 or higher.
 
 ## 1.0.0 - 2017-01-30
+
 * Compatible only with PHP 7.0 or higher. In the next versions, the library will be modified to make it compatible with PHP 5.6 or higher.
 
 ## 1.0.0 - 2016-12-14
+
 * Added `Josantonius\ErrorHandler\ErrorHandler` class.
 * Added `Josantonius\ErrorHandler\ErrorHandler->__construct()` method.
 * Added `Josantonius\ErrorHandler\ErrorHandler->catchThrowable()` method.
@@ -44,11 +90,13 @@
 * Added `Josantonius\ErrorHandler\ErrorHandler->show()` method.
 
 ## 1.0.0 - 2016-12-14
+
 * Added `Josantonius\ErrorHandler\Exception\ErrorHandlerException` class.
 * Added `Josantonius\ErrorHandler\Exception\Exceptions` abstract class.
 * Added `Josantonius\ErrorHandler\Exception\ErrorHandlerException->__construct()` method.
 
 ## 1.0.0 - 2016-12-14
+
 * Added `Josantonius\ErrorHandler\Tests\ErrorHandlerTest` class.
 * Added `Josantonius\ErrorHandler\Tests\ErrorHandlerTest->testSProvokeException()` method.
 * Added `Josantonius\ErrorHandler\Tests\ErrorHandlerTest->testSProvokeWarning()` method.